Stop removing groups before update package
The group associations to a package were removed is for maintaining the
consistency between the updated entry information and the groups. For
example if user change the product in the update then the old product
should be removed from the package. However, this will cause the flag to
be removed from the package. More complex logic can be devised to solve
the problem but it may introduce more buggy code and will lead to more
SQL calls and may even make things more confusing. Thus at the current
stage, it is better to stop removing group information.